Move awscli use over to CopyOptimizer
